safepoints and stuff
- added bank
- added teleporting
- added hella stuff but forgor

+++ Robot Game/Assets/Scripts/ItemScripts/Item.cs
@@ -1,13 +1,14 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using System.Numerics;
 
 public class Item
 {
     public int itemID;
-    public int quanity;
+    public BigInteger quanity;
 
-    public Item(int itemID, int quanity)
+    public Item(int itemID, BigInteger quanity)
     {
         this.itemID = itemID;
         this.quanity = quanity;
